+++ src/main/java/fr/hb/picom/service/StopService.java
@@ -5,7 +5,7 @@ import fr.hb.picom.business.Stop;
 import java.util.List;
 
 public interface StopService {
-    Stop getStop(Long id);
+    Stop getStop(Integer id);
     List<Stop> getStops();
 
     Stop addStop(Stop stop);
